generateur_cloud/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
pandas
numpy
Faker
azure-storage-blob
python-dotenv
tenacity
